+++ BootServer/net_channel_redis.h
@@ -12,7 +12,7 @@ extern "C" {
 #endif
 
 __declspec_dll NetChannel_t* openNetChannelRedisClient(const char* ip, unsigned short port, FnChannelRedisOnSubscribe_t on_subscribe, void* sche);
-__declspec_dll void sendRedisCmdByNetChannel(NetChannel_t* channel, int rpc_id, const char* format, ...);
+__declspec_dll void sendRedisCmdByNetChannel(NetChannel_t* channel, int64_t rpc_id, const char* format, ...);
 
 #ifdef __cplusplus
 }
